move token back to modals
