+++ src/packages/documents/documents/modals/index.ts
@@ -4,4 +4,5 @@ export * from './save-modal/index.js';
 export * from './unpublish-modal/index.js';
 export * from './schedule-modal/index.js';
 export * from './shared/index.js';
+export * from '../modals/document-picker-modal.token.js';
 export type * from './types.js';
