--- src/packages/documents/documents/picker/index.ts
@@ -1,1 +1,0 @@
-export * from './document-picker-modal.token.js';
