+++ src/packages/documents/documents/recycle-bin/entity-action/manifests.ts
@@ -2,7 +2,7 @@ import { UMB_DOCUMENT_RECYCLE_BIN_REPOSITORY_ALIAS } from '../repository/index.j
 import { UMB_DOCUMENT_ENTITY_TYPE } from '../../entity.js';
 import { UMB_DOCUMENT_ITEM_REPOSITORY_ALIAS } from '../../repository/index.js';
 import { UMB_DOCUMENT_RECYCLE_BIN_ROOT_ENTITY_TYPE } from '../entity.js';
-import { UMB_DOCUMENT_PICKER_MODAL } from '../../picker/index.js';
+import { UMB_DOCUMENT_PICKER_MODAL } from '../../modals/index.js';
 import {
 	UMB_USER_PERMISSION_DOCUMENT_DELETE,
 	UMB_USER_PERMISSION_DOCUMENT_MOVE,
