+++ src/packages/documents/documents/user-permissions/input-document-granular-user-permission/input-document-granular-user-permission.element.ts
@@ -1,6 +1,6 @@
 import type { UmbDocumentUserPermissionModel } from '../types.js';
 import { UmbDocumentItemRepository, type UmbDocumentItemModel } from '../../repository/index.js';
-import { UMB_DOCUMENT_PICKER_MODAL } from '../../picker/index.js';
+import { UMB_DOCUMENT_PICKER_MODAL } from '../../modals/index.js';
 import { css, customElement, html, repeat, state } from '@umbraco-cms/backoffice/external/lit';
 import { UmbLitElement } from '@umbraco-cms/backoffice/lit-element';
 import type { UmbModalManagerContext } from '@umbraco-cms/backoffice/modal';
